refactor(server): move db/seal/walrus/sui into src/storage/ (ENG-1747 phase 4.1)
Groups the low-level persistence + chain modules under src/storage/:
git mv src/{db,seal,walrus,sui}.rs -> src/storage/, + src/storage/mod.rs
(pub mod db/seal/sui/walrus + module doc). Pure relocation — no logic
change; git tracks all four as renames.

Plumbing updates:
- main.rs: mod db/seal/sui/walrus -> mod storage; use db::VectorDb ->
  use storage::db::VectorDb.
- crate::db / crate::seal / crate::walrus / crate::sui references ->
  crate::storage::* in types.rs, auth.rs, engine/{plaintext,walrus_seal}.rs,
  routes.rs, jobs.rs.
- src/storage/db.rs: include_str!("../migrations/...") ->
  include_str!("../../migrations/...") (one extra level deep now), x7.

cargo check + cargo test (135 passed) clean — zero warnings.

Refs ENG-1747
